refactor: 투표 조회- 데이터 파일명 변경 및 Import 경로 수정
