+++ src/features/vote/model/responseBody.ts
